refactor(tests): reorganize speech synthesis tests into dedicated file
- Moved speech synthesis tests from direct.test.ts to a new speech.test.ts file for improved organization and maintainability.
- Enhanced test clarity by utilizing the buildAgentRuntimeMock for agent runtime setup.
- Removed outdated tests from direct.test.ts, adhering to the single-responsibility principle.
